+++ gui/Constants/MyFont.java
@@ -1,0 +1,7 @@
+package Constants;
+
+import java.awt.*;
+
+public interface MyFont {
+     static Font Monospaced = new Font(Font.MONOSPACED,Font.PLAIN,12);
+}
