+++ gui/Constants/MyHeightWidth.java
@@ -1,0 +1,6 @@
+package Constants;
+
+public class MyHeightWidth {
+    public static int W = 650;
+    public static int H = 600;
+}
