--- gui/Controller/My_Font.java
@@ -1,7 +1,0 @@
-package Controller;
-
-import java.awt.*;
-
-public class My_Font {
-     static Font My_Font = new Font(Font.MONOSPACED,Font.PLAIN,12);
-}
