--- gui/Controller/Start_Event.java
@@ -1,14 +1,0 @@
-package Controller;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class Start_Event implements ActionListener {
-    public Frame obj;
-    public Start_Event(Frame obj) {
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-    }
-}
