+++ gui/Controller/panel/Menu_bar.java
@@ -6,5 +6,5 @@ public interface Menu_bar {
     public static JMenu Menu = new JMenu("Help");
     public static JMenuItem Item_AboutUs = new JMenuItem("About us");
 
-     JMenuBar add();
+    JMenuBar add();
 }
