+++ gui/Controller/panel/panel_nishi.java
@@ -1,7 +1,6 @@
 package Controller.panel;
 
 import javax.swing.*;
-import java.awt.*;
 
 public interface panel_nishi {
      JPanel panel_nishi();
