--- gui/Controller/thong_so.java
@@ -1,6 +1,0 @@
-package Controller;
-
-public interface thong_so {
-    public static int W = 650;
-    public static int H = 600;
-}
