--- gui/Event/Event_popup_frameHistory.java
@@ -1,13 +1,0 @@
-package Event;
-
-import javax.swing.*;
-import java.awt.event.ActionEvent;
-
-public class Event_popup_frameHistory extends AbstractAction {
-    JFrame frame_history = new JFrame("History");
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        
-    }
-}
