+++ gui/Event/StartEvent.java
@@ -1,0 +1,16 @@
+package Event;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+import Controller.Frame;
+
+public class StartEvent implements ActionListener {
+    public Frame obj;
+    public StartEvent(Frame obj) {
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+    }
+}
