--- gui/java_package_testing/Main1.java
@@ -1,9 +1,0 @@
-import packages.one.A;
-//import packages.two.A;
-
-public class Main1 {
-    public static void main (String[] args) {
-        A a = new A();
-        System.out.println(a.msg);
-    }
-}
