--- gui/java_package_testing/packages/one/A.java
@@ -1,6 +1,0 @@
-package packages.one;
-
-public class A {
-    public String msg = null;
-    public A() { msg = "This is from one.A"; }
-}
