--- gui/java_package_testing/packages/two/A.java
@@ -1,6 +1,0 @@
-package packages.two;
-
-public class A {
-    public String msg = null;
-    public A() { msg = "This is from two.A"; }
-}
