+++ gui/script/s.sh
@@ -1,6 +1,7 @@
 javac -encoding utf-8 -d bin ComponentDesign/*.java
+javac -encoding utf-8 -d bin Constants/*.java
 javac -encoding utf-8 -d bin Controller/*.java
 javac -encoding utf-8 -d bin Controller/panel/*.java
 javac -encoding utf-8 -d bin Event/*.java
-javac -encoding utf-8 -d bin *.java
+javac -encoding utf-8 -d bin ./*.java
 java -cp bin main_body
